+++ integration/common/openlineage/common/provider/dbt/utils.py
@@ -1,0 +1,3 @@
+__version__ = "1.25.0"
+
+PRODUCER = f"https://github.com/OpenLineage/OpenLineage/tree/{__version__}/integration/dbt"
